feat(fat0): Remove fat0.State and blockheight validation for fat0.Transaction
A fat0.State object does not belong in this package since state is
inherently tied with the database and the state engine.

A declared blockheight is not needed to help reduce lookup complexity
for tx signature uniqueness. Instead the entry hash will serve this
role. This change also means that additional external IDs on entries are
no longer allowed for fat0.Transaction entries as that would open a
replay attack up.