fix(fat0): Add timestamp to sigs to prevent replays of txs that were invalid due to insufficient balance
Previously, if a fat0.Transaction is invalid solely due to insufficient
balance, then it could be replayed at any point in the future when the
balance was sufficient. We add a timestamp to the ExtIDs to salt the
signatures. The timestamp must fall within +/- 12 hours of the timestamp
of the entry containing the fat0.Transaction. This causes signatures to
expire after some time, which is controllable by how far in the past the
timestamp salt is chosen.

Since the timestamp provides a sufficient number of options for a salt,
the fat0.Transaction.Salt field was removed.

Additionally all ExtID validation, which is shared by the fat0 types, is
now deduplicated entirely by moving those member functions to the Entry
type that Transaction and Issuance both embed.