feat(fat0): Switch back to object format for inputs/outputs, error on duplicate fields or appended content data
Previously we switched back to an array of {"address": "FA...",
"amount": 5} objects for inputs and outputs in order to avoid issues
with duplicate addresses. However this did not really eliminate the
problem entirely because any field could appear twice in a valid JSON
and this leads to ambiguity and implementation specificity. Now we
compute the expected length on the compacted JSON and compare it with
the compacted JSON in the entry content. This catches any errors due to
duplicate fields or additional data appeneded to the JSON in the entry
content. For example, previously the content "{ <valid fat0 entry> }{}"
would have been valid despite the additional "{}". Although this
particular data would make no difference, it is simpler if any entry
that is not just a single JSON object, or is a JSON object with
duplicate fields is rejected by default.