feat(fat0,fat1): Add complete fat1.Transaction validation, refactor fat0 types
Significant refactor of fat0 and fat1 type methods. The biggest refactor
is that now all types check their own JSON length and perform complete
data validation during unmarshaling and marshaling. This makes it
impossible to convert to or from JSON with an invalid type. Addtionally
fat0.Entry has been improved to be more useful when embedded in a type
in another package, like fat1.