feat(fat): Remove Issuance.Name to reduce confusion with Token ID
Token Chains have a Token ID used in the derivation of the Chain ID.
This Token ID can be any valid unicode string. Additionally the Issuance
Entry has a Name field that allows for a more descriptive name for the
token. After some discussions it was decided that this is redundant and
confusing since the Token ID will likely be used as the proper name of
the token in question. If an Issuer wishes to add additional information
about the token to the Issuance Entry they still may do so using the
Metadata field.

BREAKING CHANGE: Any Token Chains previously issued using the Name field
will now be considered invalid due to the strict parsing of the JSON
data that fatd uses. If the Token Chain did not use the Name field then
it will still be valid. Currently, fatd will not validate existing
databases. This means that if a database was built with a previous
version of fatd, it will continue parsing additional entries fro that
chain as needed, however if it attempts to rebuild the chain's fatd db
then fatd will refuse to do so. Users are advised against using
databases built with a previous version of fatd unless they really know
what they are doing. Instead rebuild all databases by removing them and
restarting fatd.