feat(srv): Query transactions for a set of addresses instead of a single address
For the 'get-transactions' method, replace the "address" field with the
"addresses" field which expects an array of strings that are human
readable factoid addresses. An empty array is invalid, simply omit the
field if you do not wish to filter transactions by address.

BREAKING CHANGE: The "address" field is deprecated. Use "addresses"
instead.

Implement #13