fix(engine): Use worker pool for processing EBlocks
Previously, a DBlock with a large number of EBlocks would launch that
number of goroutines and eventually cause an error opening new network
connections. Now we use a worker pool of a max of 16 workers, or the
number of EBlocks, whichever is fewer.

This fixes an issue reported by Steven M of TFA on the FAT Discord.