refactor(main,engine,srv): Use stop/done channels to ensure clean exit
Use a more consistent, reliable, and idiomatic paradigm for stopping on
SIGINT or errors. Now any runtime, a package that launches its own
goroutines (e.g. engine and srv), will implement the following:

func Start(stop <-struct{}) (done <-struct{})

Start should launch any goroutines and return a channel that must be
closed when all goroutines exit. The package must cleanly and safely
exit any running goroutines if stop is closed. If the done channel is
closed by the runtime before the caller of Start closes the stop
channel, then a fatal error must have occurred, and the caller should
probably close all other runtimes and exit as well. Runtimes must log
their own error with their logger, which indicates the package where the
error occurred.

The main program initializes a stop channel for each runtime and ensures
that it is closed when main exits. After all initialization, main simply
waits for SIGINT, or any runtime's done channel to be closed.

If SIGINT is received, the user has requested shutdown, so we return,
closing all stop channels and wait for all done channels to be closed.
Exit code 0.

If any done channels are closed, this indicates an error occurred, so we
exit with code 1, and also close the stop channels and wait for all done
channels to be closed.

While finally waiting for all done channels to be closed, we do not
intercept SIGINT. This way if the program hangs, the user should be able
to send a second SIGINT to forcibly exit. With the current design of
engine and srv this should never happen.