fix(engine): Allow -startscanheight 0, detect main/testnet from height
Previously -startscanheight 0 would cause an error due to some faulty
logic in the scan height checks. These checks have been moved to the
engine and have been corrected. Now 0 is an acceptable -startscanheight.
Additionally, if there is no existing database, and no -startscanheight
is given, the height of the currently connected Factom network is used
to determine whether to use the mainnet (163180) or testnet (60000)
start heights.