feat(db): Add full database validation and tests
A FAT0 and FAT1 database can be fully verified now. Any difference in
state is reported in SQL and written out to a file. Go generate is used
to create test databases, which are now checked into the repo. Running
the validation test will reveal any differences in building the database
as a result of future code changes.