refactor(db): Improved org and fewer exported funcs
Any function that modifies the database is now a member function on
*Chain. Fewer functions are exported now to reduce the API and reduce
the chance for misuse of functions that should only be called in the
context of Chain.Apply(). Additionally functions that run UPDATE are now
more consistently named Set...(). Many top level comments were removed
as a reminder to re-write the docs later.

Finally a small modification of the order of closing db connections
results in the shm and wal files being automatically removed.