refactor(engine): Replace use of state with db package
This completely removes the use of the state package from the engine
package. All of the database operations for applying new entries are in
the db package. All remaining state functionality, like processing
eblocks and passing them to the appropriate chain database is handled
entirely by the engine package.

Whitelisting and blacklisting chains is now available. Setting the
network ID is possible and is checked against the Factom blockchain.
Also users can -ignorenewchains to avoid tracking new FAT chains.