feat(srv,engine,db): Implement 'send-transaction' method
Finally you can use the 'send-transaction' RPC method. This also
addresses a potential race condition between the srv and engine
packages. We again go back to *not* using a pointer to the db.Chain.
Rollbacks are easier to understand and also encompass the Chain struct
itself and not just the database.