feat(engine,db,srv): Add pending transaction tracking support
The long awaited pending transaction support is here. This
implementation fully leverages the existing validation and database
code. No new tables, nor schema changes were required. When pending
transactions are enabled (default), when a pending transaction for a
tracked change is detected, a new in memory copy of that chains database
is created. This in memory database uses a shared cache so a pool of
read only connections can also be opened to it. A session is used to
track the changes to the in memory database as pending entries are
applied to it. When the next EBlock is published, the inverse of the
session's changeset is used to rollback all pending txs on the in memory
db. A new session on the main database is used to generate a changeset
for applying this EBlock. This session is kept open until another
pending entry is available for that database, at which point this
session's changeset is used to sync the in-memory database with the main
database. One advantage to this approach is that is allows for full code
reuse of the validation code. The downside is that a full copy of each
chain database must be kept around in memory. The -disablepending flag
has been added to allow resource constrained systems to not track
pending transactions.

re #14 #8