fix(db): Allow send to self txs in address_transaction schema
Since transactions may now have overlapping inputs and outputs, there
can be two rows in the "address_transaction" table with the same
"address_id" and "entry_id" but different values for "to". This causes a
PRIMARY_KEY constraint violation. Include "to" into the composite
primary key to resolve this.

BREAKING CHANGE: This changes the schema and will break existing
databases.

fix #49