feat(state,db): Add ContractResolver, Use temp db for state changes
Add state.ContractResolver

Previously the contract database was attached to all database
connections. Unfortunately this database sharing, even if read only,
causes database locking to a single chain.Conn. More critically, any
other connection with a read transaction open, waiting to start a write,
would need to obtain a new transaction and re-read data after another
connection had finished writing. Now the contract database is managed
through separate connections, and is not attached to all chain
databases. Now a ContractResolver maintains a read only pool of
connections and a single write connections. NumCPU goroutines are
dedicated to handling read requests of the contracts database. If a
contract database lookup fails to resolve, the request is passed to one
of NumCPU downloader goroutines which download the contract chain, and
attempt to compile it. This is returned to the requester, and also
forwarded to the single writer goroutine that will save the contract to
the contracts database. The state.State is responsible for the
ContractResolver.

Use SQLite temp database for state changes until committed

Since SQLite database connections and multi-reader/single-writer, the
previous design doesn't allow for a server thread from processing a
transaction to determine if it is valid. Instead the "send-transaction"
method has separate code to check the validity of a transaction. This is
not ideal as duplicate code, yet slightly different code, must be
maintained. With smart contracts the only way to determine the validity
of a transaction is to run/apply it. Ideally, a server thread should be
able to use the same code as the actual state engine for processing
transactions, but have all of those changes remain local to that one
server thread. This commit refactors all SQL code in the database
drivers to only make changes to the temp database and select from a
union of the temp and main databases such that the data in the temp
database takes precedence. These changes are only visible to the one
connection, and only a read transaction need be opened. A write
transaction is not opened, and the changes are not made permanent or
visible outside the connection, until a Commit function is called which
copies the data from the temp database to the main database.

There can still only be a single writer that calls commit, but that's
okay because server threads never should commit their changes to state.